Some changes

+++ ASP_PR_GuestsBook/GuestsBook/Controllers/MessagesController.cs
@@ -3,6 +3,7 @@ using GuestsBook.Repos;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using NuGet.Protocol;
 
 namespace GuestsBook.Controllers
 {
@@ -12,11 +13,19 @@ namespace GuestsBook.Controllers
         public MessagesController(IMessagesRepository repository) => _repository = repository;
 
         // GET: Messages
-        public async Task<IActionResult> Index() => View(await _repository.GetAllMessages());
+        public IActionResult Index() => View();
+
+        [HttpGet]
+        public async Task<IActionResult> GetAllMessages()
+        {
+            return ((await _repository.GetAllMessages()).Count == 0)
+                   ? Problem("No Messages")
+                   : Json((await _repository.GetAllMessages()).ToJson());
+        }
 
         public async Task<IActionResult> MyMessages()
         {
-            User userSpecification = await _repository.GetUserByLogin(HttpContext.Session.GetString("Login")).FirstAsync();
+            User userSpecification = await _repository.GetUserByLogin(HttpContext.Request.Cookies["Login"]).FirstAsync();
             return View(await _repository.GetSelfMessages(userSpecification));
         }
 
@@ -24,21 +33,13 @@ namespace GuestsBook.Controllers
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
-                return NotFound();
+                return Problem("Message was not found!");
 
             Message? msg = await _repository.GetMessageDetails(id);
 
-            if (msg == null)
-                return NotFound();
-
-            return View(msg);
-        }
-
-        // GET: Messages/Create
-        public IActionResult Create()
-        {
-            ViewData["UserId"] = new SelectList(_repository.GetUserByLogin(HttpContext.Session.GetString("Login")), "Id", "Login");
-            return View();
+            return (msg == null)
+                   ? Problem("Message was not found!")
+                   : Json(msg.ToJson());
         }
 
         // POST: Messages/Create
@@ -50,82 +51,38 @@ namespace GuestsBook.Controllers
             {
                 await _repository.CreateMessage(message);
                 await _repository.SaveChanges();
-                return RedirectToAction(nameof(MyMessages));
+                return Json("Message was added sucessfully!");
             }
-            ViewData["UserId"] = new SelectList(_repository.GetUserByLogin(HttpContext.Session.GetString("Login")), "Id", "Login", message.UserId);
-            return View(message);
+            ViewData["UserId"] = new SelectList(_repository.GetUserByLogin(HttpContext.Request.Cookies["Login"]), "Id", "Login", message.UserId).First();
+            return Problem("Add message error! Try again!");
         }
 
-        // GET: Messages/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-                return NotFound();
-
-            var message = await _repository.GetMessage(id);
-
-            if (message == null)
-                return NotFound();
-
-            ViewData["UserId"] = new SelectList(_repository.GetUserByLogin(HttpContext.Session.GetString("Login")), "Id", "Login", message.UserId);
-
-            return View(message);
-        }
-
-        // POST: Messages/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        // PUT: Messages/Edit/5
+        [HttpPut]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Mesage,UserId")] Message message)
+        public async Task<IActionResult> Edit([Bind("Id,Mesage,UserId")] Message message)
         {
-            if (id != message.Id)
-                return NotFound();
 
             if (ModelState.IsValid)
             {
-                try
-                {
-                    _repository.UpdateMessage(message);
-                    await _repository.SaveChanges();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!await _repository.MessageExists(message.Id))
-                        return NotFound();
-                    else
-                        throw;
-                }
-                return RedirectToAction(nameof(MyMessages));
+                _repository.UpdateMessage(message);
+                await _repository.SaveChanges();
+                return Json("Message was updated sucessfully!");
             }
-            ViewData["UserId"] = new SelectList(_repository.GetUserByLogin(HttpContext.Session.GetString("Login")), "Id", "Login", message.UserId);
-            return View(message);
+            return Problem("Message update error! Try again");
         }
 
-        // GET: Messages/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-                return NotFound();
 
-            var message = await _repository.GetMessageDetails(id);
-
-            if (message == null)
-                return NotFound();
-
-            return View(message);
-        }
-
-        // POST: Messages/Delete/5
-        [HttpPost, ActionName("Delete")]
+        // DELETE: Messages/Delete/5
+        [HttpDelete]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            if ((await _repository.GetAllMessages()).Count == 0)
+                return Problem("The messages list is empty! Nothing to delete!");
             await _repository.DeleteMessage(id);
             await _repository.SaveChanges();
-            return RedirectToAction(nameof(MyMessages));
+            return Json("Message was sucessfully deleted!");
         }
-
-
     }
 }
